# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "GenerateRequest": {
        "additionalProperties": false,
        "properties": {
          "inputs": {
            "title": "Inputs",
            "type": "object"
          },
          "template_type": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Template Type",
            "type": "string"
          }
        },
        "required": [
          "template_type"
        ],
        "title": "GenerateRequest",
        "type": "object"
      },
      "GenerateResponse": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "title": "GenerateResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ItemLinkResolved": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "from_item_id": {
            "title": "From Item Id",
            "type": "string"
          },
          "link_id": {
            "title": "Link Id",
            "type": "string"
          },
          "link_type": {
            "title": "Link Type",
            "type": "string"
          },
          "other_item": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ItemSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "to_item_id": {
            "title": "To Item Id",
            "type": "string"
          }
        },
        "required": [
          "link_id",
          "from_item_id",
          "to_item_id",
          "link_type",
          "created_at"
        ],
        "title": "ItemLinkResolved",
        "type": "object"
      },
      "ItemLinksResponse": {
        "additionalProperties": false,
        "properties": {
          "item_id": {
            "title": "Item Id",
            "type": "string"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/ItemLinkResolved"
            },
            "title": "Links",
            "type": "array"
          }
        },
        "required": [
          "item_id"
        ],
        "title": "ItemLinksResponse",
        "type": "object"
      },
      "ItemSummary": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "default": "",
            "title": "Created At",
            "type": "string"
          },
          "item_id": {
            "title": "Item Id",
            "type": "string"
          },
          "item_type": {
            "title": "Item Type",
            "type": "string"
          },
          "source_ref": {
            "default": "",
            "title": "Source Ref",
            "type": "string"
          },
          "source_type": {
            "default": "",
            "title": "Source Type",
            "type": "string"
          },
          "status": {
            "default": "",
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "default": "",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "item_id",
          "item_type",
          "title"
        ],
        "title": "ItemSummary",
        "type": "object"
      },
...
      "LoginRequest": {
        "additionalProperties": false,
        "properties": {
          "password": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "user_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "LoginResponse": {
        "additionalProperties": false,
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token"
        ],
        "title": "LoginResponse",
        "type": "object"
      },
      "MeResponse": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "role",
          "display_name"
        ],
        "title": "MeResponse",
        "type": "object"
      },
      "MessageResponse": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "type": "object"
      },
...
      "QARequest": {
        "additionalProperties": false,
        "properties": {
          "question": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "QARequest",
        "type": "object"
      },
      "QAResponse": {
        "additionalProperties": false,
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "answer"
        ],
        "title": "QAResponse",
        "type": "object"
      },
      "ResolveItemsRequest": {
        "additionalProperties": false,
        "properties": {
          "item_ids": {
            "items": {
              "type": "string"
            },
            "title": "Item Ids",
            "type": "array"
          }
        },
        "title": "ResolveItemsRequest",
        "type": "object"
      },
      "ResolveItemsResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ItemSummary"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "title": "ResolveItemsResponse",
        "type": "object"
      },
      "SavedPromptCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "maxLength": 12000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "tags": {
            "default": "",
            "maxLength": 2000,
            "title": "Tags",
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "content"
        ],
        "title": "SavedPromptCreateRequest",
        "type": "object"
      },
      "SavedPromptPageResponse": {
        "additionalProperties": false,
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SavedPromptResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "limit",
          "offset",
          "has_more"
        ],
        "title": "SavedPromptPageResponse",
        "type": "object"
      },
      "SavedPromptResponse": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "index_error": {
            "default": "",
            "title": "Index Error",
            "type": "string"
          },
          "index_status": {
            "default": "pending",
            "enum": [
              "pending",
              "indexed",
              "failed",
              "unavailable",
              "excluded"
            ],
            "title": "Index Status",
            "type": "string"
          },
          "tags": {
            "title": "Tags",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "content",
          "tags",
          "created_at",
          "updated_at"
        ],
        "title": "SavedPromptResponse",
        "type": "object"
      },
...
      "Source": {
        "additionalProperties": false,
        "properties": {
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "snippet": {
            "title": "Snippet",
            "type": "string"
          },
          "source_type": {
            "enum": [
              "knowledge",
              "logbook",
              "prompt",
              "document",
              "photo"
            ],
            "title": "Source Type",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "source_type",
          "title",
          "snippet"
        ],
        "title": "Source",
        "type": "object"
      },
      "TemplateMetaItem": {
        "additionalProperties": false,
        "properties": {
          "fields": {
            "items": {
              "type": "string"
            },
            "title": "Fields",
            "type": "array"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "value",
          "label"
        ],
        "title": "TemplateMetaItem",
        "type": "object"
      },
      "TemplatesMetaResponse": {
        "additionalProperties": false,
        "properties": {
          "templates": {
            "items": {
              "$ref": "#/components/schemas/TemplateMetaItem"
            },
            "title": "Templates",
            "type": "array"
          }
        },
        "title": "TemplatesMetaResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Knowledge Workspace API",
    "version": "5.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/generate": {
      "post": {
        "operationId": "generate_api_generate_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate"
      }
    },
...
    "/api/item-links": {
      "get": {
        "operationId": "list_item_links_api_item_links_get",
        "parameters": [
          {
            "in": "query",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemLinksResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Item Links"
      }
    },
    "/api/items/resolve": {
      "post": {
        "operationId": "resolve_items_api_items_resolve_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveItemsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveItemsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Items"
      }
    },
...
    "/api/login": {
      "post": {
        "operationId": "login_api_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login"
      }
    },
    "/api/me": {
      "get": {
        "operationId": "me_api_me_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Me"
      }
    },
    "/api/meta/templates": {
      "get": {
        "operationId": "list_templates_api_meta_templates_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplatesMetaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Templates"
      }
    },
...
    "/api/prompts": {
      "get": {
        "operationId": "list_saved_prompts_api_prompts_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedPromptPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Saved Prompts"
      },
      "post": {
        "operationId": "create_saved_prompt_api_prompts_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedPromptCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedPromptResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Saved Prompt"
      }
    },
    "/api/prompts/{prompt_id}": {
      "delete": {
        "operationId": "delete_saved_prompt_api_prompts__prompt_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "prompt_id",
            "required": true,
            "schema": {
              "title": "Prompt Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Saved Prompt"
      }
    },
    "/api/qa": {
      "post": {
        "operationId": "qa_api_qa_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QARequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QAResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Qa"
      }
    },
    "/api/search": {
      "get": {
        "description": "Global search across workspace entities.\n\nQuery params:\n- q: keyword substring match\n- types: comma-separated list: knowledge,logbook,document,photo,prompt,autotest_run\n  Unsupported values return HTTP 400.\n- status_filter: exact status match (e.g. draft/reviewed/verified/archived/passed/failed)\n- tag: tags substring match (where applicable)\n- date_from/date_to: ISO comparisons; a date-only date_to includes the full day (e.g. 2026-04-01)",
        "operationId": "global_search_api_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "types",
            "required": false,
            "schema": {
              "default": "",
              "title": "Types",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status_filter",
            "required": false,
            "schema": {
              "default": "",
              "title": "Status Filter",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "default": "",
              "title": "Tag",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "default": "",
              "title": "Date From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "default": "",
              "title": "Date To",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveItemsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Global Search"
      }
    },
...
  }
}